* Register EmailTemplateManager as an osgi service. * remove config package and move service class.
