+++ components/email-mgt/org.wso2.carbon.email.mgt/src/main/java/org/wso2/carbon/email/mgt/EmailTemplateManagerImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.wso2.carbon.email.mgt.config;
+package org.wso2.carbon.email.mgt;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
