+++ components/email-mgt/org.wso2.carbon.email.mgt/src/main/java/org/wso2/carbon/email/mgt/internal/TenantManagementListener.java
@@ -20,8 +20,8 @@ package org.wso2.carbon.email.mgt.internal;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.email.mgt.config.EmailTemplateManager;
-import org.wso2.carbon.email.mgt.config.EmailTemplateManagerImpl;
+import org.wso2.carbon.email.mgt.EmailTemplateManager;
+import org.wso2.carbon.email.mgt.EmailTemplateManagerImpl;
 import org.wso2.carbon.email.mgt.exceptions.I18nEmailMgtException;
 import org.wso2.carbon.stratos.common.beans.TenantInfoBean;
 import org.wso2.carbon.stratos.common.exception.StratosException;
